Add files via upload
This program asks you for what to replace and what do replace it with. Then, the program replaces what you told it to replace with what you told it to replace it with. For example, If you were writing an essay for your client John, but you thought that his name was Jack, so you wrote Jack instead of John every time you mentioned his name in the essay. Now, if you don't want to delete every Jack and replace it with a John, you can use this program by typing in Jack (the value getting replaced) and when it asks you for another input, enter John (the value that is replacing Jack). Of course, that is just an example because this helpful tool can help you save a few moments of time, but the best part of this is seeing how it will replace every instance of one thing with another. Just like magic!!